configs: .profile -> profile.sh
Use "profile.sh" for the shared setup of zsh and bash interactive login
shells, and source it from .bash_profile and .zprofile.

The name ".profile" implies that it's POSIX compliant and should be
linked to $HOME, neither of which are true (we don't want POSIX shells
like sh trying to load it). Let's add a .bash_profile and link that to
$HOME instead. This makes it clear that it's a bash profile and is the
first profile path checked by bash for login shells, so it won't be
overridden by other files due to precedence.